+++ tests/testthat/test-timeless.R
@@ -11,3 +11,8 @@ test_that("chronos works with supported formats", {
 
     expect_true(!any(is.na(chronos(bench_date))))
 })
+
+test_that("timezone conversion #14", {
+    res <- chronos(bench_date[5], tz = "America/Los_Angeles", to_tz = "CET")
+    expect_true(res == as.POSIXct("2017-11-26 08:34:50 CET"))
+})